Set indexType for IndexBuffer only VK_INDEX_TYPE_UINT32

+++ Source/Graphics/Mesh.cpp
@@ -7,6 +7,6 @@ void Mesh::setVertices(const std::vector<float> vertices) {
     this->vertices = vertices;
 }
 
-void Mesh::setIndices(const std::vector<uint16_t> indices) {
+void Mesh::setIndices(const std::vector<uint32_t> indices) {
     this->indices = indices;
 }
